Finish converting to Selenium
Will refactor now because the codegen I used made a new class for each.
