--- src/tests/test_checkBrowse.py
@@ -1,81 +1,0 @@
-# Generated by Selenium IDE
-import pytest
-import time
-import json
-from selenium import webdriver
-from selenium.webdriver.common.by import By
-from selenium.webdriver.common.action_chains import ActionChains
-from selenium.webdriver.support import expected_conditions
-from selenium.webdriver.support.wait import WebDriverWait
-from selenium.webdriver.common.keys import Keys
-from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
-
-
-class TestCheckBrowse:
-    def setup_method(self, method):
-        self.driver = webdriver.Chrome()
-        self.vars = {}
-
-    def teardown_method(self, method):
-        self.driver.quit()
-
-    def test_checkBrowse(self):
-        self.driver.get("http://127.0.0.1:5000")
-        self.driver.set_window_size(1245, 1040)
-        self.driver.find_element(By.CSS_SELECTOR, ".centered-content").click()
-        self.driver.find_element(By.LINK_TEXT, "Browse Posts").click()
-        assert (
-            self.driver.find_element(
-                By.CSS_SELECTOR, ".card:nth-child(4) .card-title"
-            ).text
-            == "Alice\\'s Post #1"
-        )
-        assert (
-            self.driver.find_element(
-                By.CSS_SELECTOR, ".card:nth-child(6) .card-title"
-            ).text
-            == "Alice\\'s Post #2"
-        )
-        self.driver.execute_script("window.scrollTo(0,800)")
-        assert (
-            self.driver.find_element(
-                By.CSS_SELECTOR, ".card:nth-child(8) .card-title"
-            ).text
-            == "Alice\\'s Post #3"
-        )
-        assert (
-            self.driver.find_element(
-                By.CSS_SELECTOR, ".card:nth-child(10) .card-title"
-            ).text
-            == "Bob\\'s Post #1"
-        )
-        assert (
-            self.driver.find_element(
-                By.CSS_SELECTOR, ".card:nth-child(12) .card-title"
-            ).text
-            == "Bob\\'s Post #2"
-        )
-        assert (
-            self.driver.find_element(
-                By.CSS_SELECTOR, ".card:nth-child(14) .card-title"
-            ).text
-            == "Bob\\'s Post #3"
-        )
-        assert (
-            self.driver.find_element(
-                By.CSS_SELECTOR, ".card:nth-child(16) .card-title"
-            ).text
-            == "Carol\\'s Post #1"
-        )
-        assert (
-            self.driver.find_element(
-                By.CSS_SELECTOR, ".card:nth-child(18) .card-title"
-            ).text
-            == "Carol\\'s Post #2"
-        )
-        assert (
-            self.driver.find_element(
-                By.CSS_SELECTOR, ".card:nth-child(20) .card-title"
-            ).text
-            == "Carol\\'s Post #3"
-        )
