--- src/tests/test_likeButton.py
@@ -1,85 +1,0 @@
-# Generated by Selenium IDE
-import pytest
-import time
-import json
-from selenium import webdriver
-from selenium.webdriver.common.by import By
-from selenium.webdriver.common.action_chains import ActionChains
-from selenium.webdriver.support import expected_conditions
-from selenium.webdriver.support.wait import WebDriverWait
-from selenium.webdriver.common.keys import Keys
-from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
-
-
-class TestLikeButton:
-    def setup_method(self, method):
-        self.driver = webdriver.Chrome()
-        self.vars = {}
-
-    def teardown_method(self, method):
-        self.driver.quit()
-
-    def test_likeButton(self):
-        self.driver.get("http://127.0.0.1:5000")
-        self.driver.set_window_size(1227, 1020)
-        self.driver.find_element(By.LINK_TEXT, "Log In").click()
-        self.driver.find_element(By.ID, "login-email").click()
-        self.driver.find_element(By.ID, "login-email").send_keys("alice@example.com")
-        self.driver.find_element(By.ID, "login-password").send_keys("password123")
-        self.driver.find_element(By.ID, "login-password").send_keys(Keys.ENTER)
-        assert (
-            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
-            == "Successfully logged in\\\\n×"
-        )
-        self.driver.find_element(By.LINK_TEXT, "Browse Posts").click()
-        assert (
-            self.driver.find_element(
-                By.CSS_SELECTOR, ".card:nth-child(4) .d-flex > .btn-outline-primary"
-            ).text
-            == "Like (0)"
-        )
-        self.driver.find_element(
-            By.CSS_SELECTOR, ".card:nth-child(4) .d-flex > .btn-outline-primary"
-        ).click()
-        assert (
-            self.driver.find_element(
-                By.CSS_SELECTOR, ".card:nth-child(4) .d-flex > .btn-outline-primary"
-            ).text
-            == "Unlike (1)"
-        )
-        self.driver.find_element(By.LINK_TEXT, "Log Out").click()
-        assert (
-            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
-            == "You have been logged out.\\\\n×"
-        )
-        self.driver.find_element(By.LINK_TEXT, "Log In").click()
-        element = self.driver.find_element(By.LINK_TEXT, "Log In")
-        actions = ActionChains(self.driver)
-        actions.move_to_element(element).perform()
-        element = self.driver.find_element(By.CSS_SELECTOR, "body")
-        actions = ActionChains(self.driver)
-        actions.move_to_element(element, 0, 0).perform()
-        self.driver.find_element(By.ID, "login-email").click()
-        self.driver.find_element(By.ID, "login-email").send_keys("bob@example.com")
-        self.driver.find_element(By.ID, "login-password").send_keys("password123")
-        self.driver.find_element(By.ID, "login-password").send_keys(Keys.ENTER)
-        assert (
-            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
-            == "Successfully logged in\\\\n×"
-        )
-        self.driver.find_element(By.LINK_TEXT, "Browse Posts").click()
-        assert (
-            self.driver.find_element(
-                By.CSS_SELECTOR, ".card:nth-child(4) .d-flex > .btn-outline-primary"
-            ).text
-            == "Like (1)"
-        )
-        self.driver.find_element(
-            By.CSS_SELECTOR, ".card:nth-child(4) .d-flex > .btn-outline-primary"
-        ).click()
-        assert (
-            self.driver.find_element(
-                By.CSS_SELECTOR, ".card:nth-child(4) .d-flex > .btn-outline-primary"
-            ).text
-            == "Unlike (2)"
-        )
