--- src/tests/test_notificationTest.py
@@ -1,67 +1,0 @@
-# Generated by Selenium IDE
-import pytest
-import time
-import json
-from selenium import webdriver
-from selenium.webdriver.common.by import By
-from selenium.webdriver.common.action_chains import ActionChains
-from selenium.webdriver.support import expected_conditions
-from selenium.webdriver.support.wait import WebDriverWait
-from selenium.webdriver.common.keys import Keys
-from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
-
-
-class TestNotificationTest:
-    def setup_method(self, method):
-        self.driver = webdriver.Chrome()
-        self.vars = {}
-
-    def teardown_method(self, method):
-        self.driver.quit()
-
-    def test_notificationTest(self):
-        self.driver.get("http://127.0.0.1:5000")
-        self.driver.set_window_size(1285, 1039)
-        self.driver.find_element(By.LINK_TEXT, "Log In").click()
-        self.driver.find_element(By.ID, "login-email").click()
-        self.driver.find_element(By.ID, "login-email").send_keys("alice@example.com")
-        self.driver.find_element(By.ID, "login-password").send_keys("password123")
-        self.driver.find_element(By.ID, "login-password").send_keys(Keys.ENTER)
-        self.driver.find_element(By.CSS_SELECTOR, ".centered-content").click()
-        assert (
-            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
-            == "Successfully logged in\\\\n×"
-        )
-        assert (
-            self.driver.find_element(
-                By.CSS_SELECTOR, ".col-md-3:nth-child(3) .card-title"
-            ).text
-            == "Responses2"
-        )
-        self.driver.find_element(By.LINK_TEXT, "Notifications").click()
-        assert self.driver.find_element(By.LINK_TEXT, "Bob Smith").text == "Bob Smith"
-        assert (
-            self.driver.find_element(By.LINK_TEXT, "Carol Martinez").text
-            == "Carol Martinez"
-        )
-        self.driver.find_element(By.CSS_SELECTOR, ".alert:nth-child(2) span").click()
-        assert (
-            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
-            == "Notification dismissed.\\\\n×"
-        )
-        self.driver.find_element(By.CSS_SELECTOR, "form span").click()
-        assert (
-            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
-            == "Notification dismissed.\\\\n×"
-        )
-        assert (
-            self.driver.find_element(By.CSS_SELECTOR, "p").text
-            == "No notifications to show."
-        )
-        self.driver.find_element(By.LINK_TEXT, "Love Letters").click()
-        assert (
-            self.driver.find_element(
-                By.CSS_SELECTOR, ".col-md-3:nth-child(3) .card-title"
-            ).text
-            == "Responses0"
-        )
