+++ src/tests/test_selenium.py
@@ -5,7 +5,7 @@ import json
 from selenium import webdriver
 from selenium.webdriver.common.by import By
 from selenium.webdriver.common.action_chains import ActionChains
-from selenium.webdriver.support import expected_conditions
+from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.wait import WebDriverWait
 from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
@@ -31,6 +31,7 @@ class Test_login:
     def setup_method(self, method):
         self.driver = webdriver.Chrome()
         self.vars = {}
+        reset_database()
 
     def teardown_method(self, method):
         self.driver.quit()
@@ -45,15 +46,15 @@ class Test_login:
         self.driver.find_element(By.ID, "login-password").send_keys("password123")
         self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(6)").click()
 
-        self.driver.implicitly_wait(10)
+        time.sleep(0.5)
 
         self.driver.find_element(By.CSS_SELECTOR, ".alert-success").click()
-
 
 class Test_createPost:
     def setup_method(self, method):
         self.driver = webdriver.Chrome()
         self.vars = {}
+        reset_database()
 
     def teardown_method(self, method):
         self.driver.quit()
@@ -100,7 +101,6 @@ class Test_createPost:
             == "Selenium Test"
         )
 
-
 class Test_checkReplies:
     def setup_method(self, method):
         self.driver = webdriver.Chrome()
@@ -135,7 +135,6 @@ class Test_checkReplies:
             self.driver.find_element(By.CSS_SELECTOR, "#replies-3 .card-subtitle").text
             == "Reply by Anonymous"
         )
-
 
 class Test_passwordChange:
     def setup_method(self, method):
@@ -235,4 +234,648 @@ class Test_passwordChange:
         assert (
             self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
             == "Successfully logged in\n×"
+        )
+
+class Test_notificationcheck:
+    def setup_method(self, method):
+        self.driver = webdriver.Chrome()
+        self.vars = {}
+        reset_database()
+
+    def teardown_method(self, method):
+        self.driver.quit()
+        reset_database()
+
+    def test_notificationcheck(self):
+        self.driver.get("http://127.0.0.1:5000")
+        self.driver.set_window_size(1285, 1039)
+        self.driver.find_element(By.LINK_TEXT, "Log In").click()
+        self.driver.find_element(By.ID, "login-email").click()
+        self.driver.find_element(By.ID, "login-email").send_keys("alice@example.com")
+        self.driver.find_element(By.ID, "login-password").send_keys("password123")
+        self.driver.find_element(By.ID, "login-password").send_keys(Keys.ENTER)
+
+        time.sleep(0.5)
+
+        self.driver.find_element(By.CSS_SELECTOR, ".centered-content").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Successfully logged in\n×"
+        )
+        assert (
+            (self.driver.find_element(
+                By.CSS_SELECTOR, ".col-md-3:nth-child(3) .card-title"
+            ).text
+            == "Responses1") or
+            (self.driver.find_element(
+                By.CSS_SELECTOR, ".col-md-3:nth-child(3) .card-title"
+            ).text
+            == "Responses2")
+
+        )
+
+        time.sleep(0.5)
+
+        self.driver.find_element(By.LINK_TEXT, "Notifications").click()
+
+        time.sleep(0.5)
+
+        assert self.driver.find_element(By.LINK_TEXT, "Bob Smith").text == "Bob Smith"
+        assert (
+            self.driver.find_element(By.LINK_TEXT, "Carol Martinez").text
+            == "Carol Martinez"
+        )
+        self.driver.find_element(By.CSS_SELECTOR, ".alert:nth-child(2) span").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Notification dismissed.\n×"
+        )
+        self.driver.find_element(By.CSS_SELECTOR, "form span").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Notification dismissed.\n×"
+        )
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, "p").text
+            == "No notifications to show."
+        )
+        self.driver.find_element(By.LINK_TEXT, "Love Letters").click()
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".col-md-3:nth-child(3) .card-title"
+            ).text
+            == "Responses0"
+        )
+
+class Test_checkBrowse:
+    def setup_method(self, method):
+        self.driver = webdriver.Chrome()
+        self.vars = {}
+
+    def teardown_method(self, method):
+        self.driver.quit()
+        reset_database()
+
+    def test_checkBrowse(self):
+        self.driver.get("http://127.0.0.1:5000")
+        self.driver.set_window_size(1245, 1040)
+        self.driver.find_element(By.CSS_SELECTOR, ".centered-content").click()
+        self.driver.find_element(By.LINK_TEXT, "Browse Posts").click()
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".card:nth-child(4) .card-title"
+            ).text
+            == "Alice's Post #1"
+        )
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".card:nth-child(6) .card-title"
+            ).text
+            == "Alice's Post #2"
+        )
+        self.driver.execute_script("window.scrollTo(0,800)")
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".card:nth-child(8) .card-title"
+            ).text
+            == "Alice's Post #3"
+        )
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".card:nth-child(10) .card-title"
+            ).text
+            == "Bob's Post #1"
+        )
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".card:nth-child(12) .card-title"
+            ).text
+            == "Bob's Post #2"
+        )
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".card:nth-child(14) .card-title"
+            ).text
+            == "Bob's Post #3"
+        )
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".card:nth-child(16) .card-title"
+            ).text
+            == "Carol's Post #1"
+        )
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".card:nth-child(18) .card-title"
+            ).text
+            == "Carol's Post #2"
+        )
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".card:nth-child(20) .card-title"
+            ).text
+            == "Carol's Post #3"
+        )
+
+class Test_makeReplies:
+    def setup_method(self, method):
+        self.driver = webdriver.Chrome()
+        self.vars = {}
+        reset_database()
+
+    def teardown_method(self, method):
+        self.driver.quit()
+        reset_database()
+
+    def test_makeReplies(self):
+        self.driver.get("http://127.0.0.1:5000")
+        self.driver.set_window_size(1285, 1039)
+        self.driver.find_element(By.LINK_TEXT, "Log In").click()
+        self.driver.find_element(By.ID, "login-email").click()
+        self.driver.find_element(By.ID, "login-email").send_keys("alice@example.com")
+        self.driver.find_element(By.ID, "login-password").send_keys("password123")
+        self.driver.find_element(By.ID, "login-password").send_keys(Keys.ENTER)
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Successfully logged in\n×"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Browse Posts").click()
+        self.driver.find_element(
+            By.CSS_SELECTOR, ".card:nth-child(4) .d-flex > .btn-primary"
+        ).click()
+        self.driver.find_element(By.NAME, "content").click()
+        self.driver.find_element(By.NAME, "content").send_keys("Sent from Selenium")
+
+        time.sleep(0.5)
+
+        self.driver.find_element(
+            By.CSS_SELECTOR, ".modal-footer > .btn-primary"
+        ).click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, "#replies-1 > .card:nth-child(1) .card-subtitle"
+            ).text
+            == "Reply by You"
+        )
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, "#replies-1 > .card:nth-child(1) .card-text"
+            ).text
+            == "Sent from Selenium"
+        )
+
+class Test_profileEdit:
+    def setup_method(self, method):
+        self.driver = webdriver.Chrome()
+        self.vars = {}
+        reset_database()
+
+    def teardown_method(self, method):
+        self.driver.quit()
+        reset_database()
+
+    def test_profileEdit(self):
+        self.driver.get("http://127.0.0.1:5000")
+        self.driver.set_window_size(1167, 1020)
+        self.driver.find_element(By.LINK_TEXT, "Log In").click()
+        self.driver.find_element(By.ID, "login-email").click()
+        self.driver.find_element(By.ID, "login-email").send_keys("alice@example.com")
+        self.driver.find_element(By.ID, "login-password").send_keys("password123")
+        self.driver.find_element(By.ID, "login-password").send_keys(Keys.ENTER)
+
+        time.sleep(0.5)
+
+        self.driver.find_element(By.LINK_TEXT, "Alice Johnson").click()
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, "p:nth-child(5)").text
+            == "Gender: Female"
+        )
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, "p:nth-child(7)").text
+            == "Instagram: alice_j"
+        )
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, "p:nth-child(9)").text
+            == "Snapchat: Not provided"
+        )
+
+        time.sleep(0.5)
+
+        self.driver.find_element(By.ID, "editButton").click()
+
+        time.sleep(0.5)
+
+        self.driver.find_element(By.ID, "edit_instagram").clear()
+        self.driver.find_element(By.ID, "edit_instagram").click()
+        self.driver.find_element(By.ID, "edit_instagram").send_keys("alice_james")
+        self.driver.find_element(By.ID, "edit_snapchat").clear()
+        self.driver.find_element(By.ID, "edit_snapchat").click()
+        self.driver.find_element(By.ID, "edit_snapchat").send_keys("Selenium")
+
+        time.sleep(0.5)
+
+        self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(9)").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Account details successfully updated\n×"
+        )
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, "p:nth-child(7)").text
+            == "Instagram: alice_james"
+        )
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, "p:nth-child(9)").text
+            == "Snapchat: Selenium"
+        )
+
+class Test_notificationMake:
+    def setup_method(self, method):
+        self.driver = webdriver.Chrome()
+        self.vars = {}
+        reset_database()
+
+    def teardown_method(self, method):
+        self.driver.quit()
+        reset_database()
+
+    def test_notificationMake(self):
+        self.driver.get("http://127.0.0.1:5000")
+        self.driver.set_window_size(1167, 1020)
+        self.driver.find_element(By.LINK_TEXT, "Sign Up").click()
+        self.driver.find_element(By.ID, "first_name").click()
+        self.driver.find_element(By.ID, "first_name").send_keys("Selenium")
+        self.driver.find_element(By.ID, "last_name").send_keys("Test")
+        self.driver.find_element(By.ID, "gender").click()
+        dropdown = self.driver.find_element(By.ID, "gender")
+        dropdown.find_element(By.XPATH, "//option[. = 'Male']").click()
+        self.driver.find_element(By.ID, "email").click()
+        self.driver.find_element(By.ID, "email").send_keys("selenium@example.com")
+        self.driver.find_element(By.ID, "password").send_keys("Password123!")
+        self.driver.find_element(By.ID, "instagram").click()
+        self.driver.find_element(By.ID, "instagram").send_keys("selenium")
+        self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(13)").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Account successfully created\n×"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Log In").click()
+        element = self.driver.find_element(By.LINK_TEXT, "Log In")
+        actions = ActionChains(self.driver)
+        actions.move_to_element(element).perform()
+        element = self.driver.find_element(By.CSS_SELECTOR, "body")
+        actions = ActionChains(self.driver)
+        actions.move_to_element(element).perform()
+        self.driver.find_element(By.ID, "login-email").click()
+        self.driver.find_element(By.ID, "login-email").send_keys("selenium@example.com")
+        self.driver.find_element(By.ID, "login-password").send_keys("Password123!")
+        self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(6)").click()
+
+        time.sleep(0.5)
+    
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Successfully logged in\n×"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Browse Posts").click()
+        self.driver.find_element(
+            By.CSS_SELECTOR, ".card:nth-child(4) form > .btn"
+        ).click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Connect request sent.\n×"
+        )
+        self.driver.find_element(
+            By.CSS_SELECTOR, ".card:nth-child(10) form > .btn"
+        ).click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Connect request sent.\n×"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Log Out").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "You have been logged out.\n×"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Log In").click()
+        self.driver.find_element(By.ID, "login-email").click()
+        self.driver.find_element(By.ID, "login-email").send_keys("alice@example.com")
+        self.driver.find_element(By.ID, "login-password").send_keys("password123")
+        self.driver.find_element(By.ID, "login-password").send_keys(Keys.ENTER)
+        
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Successfully logged in\n×"
+        )
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".col-md-3:nth-child(3) .card-title"
+            ).text
+            == "Responses3"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Notifications").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.LINK_TEXT, "Selenium Test").text
+            == "Selenium Test"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Log Out").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "You have been logged out.\n×"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Log In").click()
+        self.driver.find_element(By.ID, "login-email").click()
+        self.driver.find_element(By.ID, "login-email").send_keys("bob@example.com")
+        self.driver.find_element(By.ID, "login-password").send_keys("password123")
+        self.driver.find_element(By.ID, "login-password").send_keys(Keys.ENTER)
+
+        time.sleep(0.5)
+
+        self.driver.find_element(By.LINK_TEXT, "Notifications").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.LINK_TEXT, "Selenium Test").text
+            == "Selenium Test"
+        )
+
+class Test_likeButton:
+    def setup_method(self, method):
+        self.driver = webdriver.Chrome()
+        self.vars = {}
+        reset_database()
+
+    def teardown_method(self, method):
+        self.driver.quit()
+        reset_database()
+
+    def test_likeButton(self):
+        self.driver.get("http://127.0.0.1:5000")
+        self.driver.set_window_size(1227, 1020)
+        self.driver.find_element(By.LINK_TEXT, "Log In").click()
+        self.driver.find_element(By.ID, "login-email").click()
+        self.driver.find_element(By.ID, "login-email").send_keys("alice@example.com")
+        self.driver.find_element(By.ID, "login-password").send_keys("password123")
+        self.driver.find_element(By.ID, "login-password").send_keys(Keys.ENTER)
+        time.sleep(0.5)
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Successfully logged in\n×"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Browse Posts").click()
+        
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".card:nth-child(4) .d-flex > .btn-outline-primary"
+            ).text
+            == "Like (0)"
+        )
+
+        time.sleep(0.5)
+
+        self.driver.find_element(
+            By.CSS_SELECTOR, ".card:nth-child(4) .d-flex > .btn-outline-primary"
+        ).click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".card:nth-child(4) .d-flex > .btn-outline-primary"
+            ).text
+            == "Unlike (1)"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Log Out").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "You have been logged out.\n×"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Log In").click()
+        element = self.driver.find_element(By.LINK_TEXT, "Log In")
+        actions = ActionChains(self.driver)
+        actions.move_to_element(element).perform()
+        element = self.driver.find_element(By.CSS_SELECTOR, "body")
+        actions = ActionChains(self.driver)
+        actions.move_to_element(element).perform()
+        self.driver.find_element(By.ID, "login-email").click()
+        self.driver.find_element(By.ID, "login-email").send_keys("bob@example.com")
+        self.driver.find_element(By.ID, "login-password").send_keys("password123")
+        self.driver.find_element(By.ID, "login-password").send_keys(Keys.ENTER)
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Successfully logged in\n×"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Browse Posts").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".card:nth-child(4) .d-flex > .btn-outline-primary"
+            ).text
+            == "Like (1)"
+        )
+        self.driver.find_element(
+            By.CSS_SELECTOR, ".card:nth-child(4) .d-flex > .btn-outline-primary"
+        ).click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".card:nth-child(4) .d-flex > .btn-outline-primary"
+            ).text
+            == "Unlike (2)"
+        )
+
+class Test_notLoggedIn:
+    def setup_method(self, method):
+        self.driver = webdriver.Chrome()
+        self.vars = {}
+        reset_database()
+
+    def teardown_method(self, method):
+        self.driver.quit()
+        reset_database()
+
+    def test_notLoggedIn(self):
+        self.driver.get("http://127.0.0.1:5000")
+        self.driver.set_window_size(1227, 1020)
+        self.driver.find_element(By.LINK_TEXT, "Create Post").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-warning").text
+            == "You must log in to access this page.\n×"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Notifications").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-warning").text
+            == "You must log in to access this page.\n×"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Browse Posts").click()
+        self.driver.find_element(
+            By.CSS_SELECTOR, ".card:nth-child(4) form > .btn"
+        ).click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-dismissible").text
+            == "You need to login to connect.\n×"
+        )
+
+class Test_accountValidation:
+    def setup_method(self, method):
+        self.driver = webdriver.Chrome()
+        self.vars = {}
+        reset_database()
+
+    def teardown_method(self, method):
+        self.driver.quit()
+        reset_database()
+
+    def test_accountValidation(self):
+        self.driver.get("http://127.0.0.1:5000")
+        self.driver.set_window_size(1357, 1020)
+        self.driver.find_element(By.LINK_TEXT, "Sign Up").click()
+        self.driver.find_element(By.ID, "first_name").click()
+        self.driver.find_element(By.ID, "first_name").send_keys("123")
+        self.driver.find_element(By.ID, "last_name").send_keys("123")
+        self.driver.find_element(By.ID, "email").click()
+        self.driver.find_element(By.ID, "email").send_keys("test")
+        self.driver.find_element(By.ID, "password").click()
+        self.driver.find_element(By.ID, "password").send_keys("1")
+        self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(13)").click()
+        self.driver.find_element(By.ID, "email").clear()
+        self.driver.find_element(By.ID, "email").send_keys("test@")
+        self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(13)").click()
+        self.driver.find_element(By.ID, "email").clear()
+        self.driver.find_element(By.ID, "email").send_keys("test@test")
+        self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(13)").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".form-group:nth-child(3) > .error-message"
+            ).text
+            == "First Name should contain only alphabetic characters."
+        )
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".form-group:nth-child(4) > .error-message"
+            ).text
+            == "Last Name should contain only alphabetic characters."
+        )
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".form-group:nth-child(6) > .error-message"
+            ).text
+            == "Invalid email format."
+        )
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".form-group:nth-child(7) > .error-message"
+            ).text
+            == "Password must be at least 8 characters long."
+        )
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".error-message:nth-child(13)"
+            ).text
+            == "At least one social media handle must be provided"
+        )
+        self.driver.find_element(By.ID, "first_name").click()
+        self.driver.find_element(By.ID, "first_name").clear()
+        self.driver.find_element(By.ID, "first_name").send_keys("a")
+        self.driver.find_element(By.ID, "last_name").click()
+        self.driver.find_element(By.ID, "last_name").clear()
+        self.driver.find_element(By.ID, "last_name").send_keys("a")
+        self.driver.find_element(By.ID, "email").click()
+        self.driver.find_element(By.ID, "email").clear()
+        self.driver.find_element(By.ID, "email").send_keys("test@test.com")
+        self.driver.find_element(By.ID, "password").click()
+        self.driver.find_element(By.ID, "password").clear()
+        self.driver.find_element(By.ID, "password").send_keys("1aaaaaaa")
+        self.driver.find_element(By.ID, "instagram").click()
+        self.driver.find_element(By.ID, "instagram").send_keys("a")
+        self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(14)").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".error-message").text
+            == "Password must contain at least one uppercase letter."
+        )
+        self.driver.find_element(By.ID, "password").click()
+        self.driver.find_element(By.ID, "password").clear()
+        self.driver.find_element(By.ID, "password").send_keys("1aaaaaaA")
+        self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(13)").click()
+        self.driver.find_element(By.ID, "password").click()
+        self.driver.find_element(By.ID, "password").clear()
+        self.driver.find_element(By.ID, "password").send_keys("1aaaaaaA!")
+        self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(13)").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Account successfully created\n×"
         )
